+++ src/graia/ariadne/message/parser/twilight.py
@@ -30,7 +30,6 @@ from typing import (
 from graia.broadcast.builtin.derive import DeriveDispatcher
 from graia.broadcast.entities.decorator import Decorator
 from graia.broadcast.entities.dispatcher import BaseDispatcher
-from graia.broadcast.exceptions import ExecutionStop
 from graia.broadcast.interfaces.decorator import DecoratorInterface
 from graia.broadcast.interfaces.dispatcher import DispatcherInterface
 from pydantic.utils import Representation
@@ -840,17 +839,21 @@ class ResultValue(Decorator):
     def __call__(self, _: Any, i: DispatcherInterface) -> Any:
         sparkle: Sparkle = i.local_storage[f"{__name__}:result"]
         res = sparkle.res.get(i.name, None)
+        if res is not None:
+            res = res.result
         if generic_isinstance(res, i.annotation):
-            return res.result if res else None
+            return res
         i.stop()
 
     @staticmethod
     async def target(i: DecoratorInterface):
         sparkle: Sparkle = i.local_storage[f"{__name__}:result"]
         res = sparkle.res.get(i.name, None)
+        if res is not None:
+            res = res.result
         if generic_isinstance(res, i.annotation):
-            return res.result if res else None
-        raise ExecutionStop
+            return res
+        i.dispatcher_interface.stop()
 
 
 class Help(Decorator, Generic[T]):